refactor(LottoAmount): validation 래퍼클래스에서 관리
